Eliminate all clang compiler warnings (677 -> 0)
Remove dead `register` keyword (271 sites); convert VLAs to std::vector/new[] (32); rewrite sprintf/vsprintf -> snprintf/vsnprintf with explicit bounds (131); istrstream -> istringstream and pragma-guard the residual ostrstream/strstreambuf sites (22); fix real defects -- %lx-on-pointer and %d-on-long format mismatches, ~0<<n shift-of-negative, and a null-reference guard (7); pass ACE includes with -isystem so its header deprecations stay quiet while -Wdeprecated-declarations stays live for our code; drop deprecated -Wl,-bind_at_load from the DARWIN link flags.

+++ src/ComUnidraw/pixelcmds.c
@@ -67,7 +67,7 @@ void PolyClipRasterCmd::Execute() {
   RasterOvComp* rastcomp = _rastcomp;
   VerticesOvComp* vertcomp = _vertcomp;
 
-  fprintf(stderr, "Rastcomp AT START is now %lx, Vertcomp is now %lx\n", rastcomp, vertcomp);
+  fprintf(stderr, "Rastcomp AT START is now %lx, Vertcomp is now %lx\n", (unsigned long)rastcomp, (unsigned long)vertcomp);
   for (sel->First(i); !sel->Done(i); sel->Next(i)) {
     if (rastcomp != nil && vertcomp != nil) { break; }
     GraphicView* view = sel->GetView(i);
@@ -88,7 +88,7 @@ void PolyClipRasterCmd::Execute() {
     
   }
 
-  fprintf(stderr, "Rastcomp is now %lx, Vertcomp is now %lx\n", rastcomp, vertcomp);
+  fprintf(stderr, "Rastcomp is now %lx, Vertcomp is now %lx\n", (unsigned long)rastcomp, (unsigned long)vertcomp);
 
   if (rastcomp != nil && vertcomp != nil) {
     _rastcomp = rastcomp;
